add imagens e ajustando textos
