+++ src/Components/Home/Home.jsx
@@ -368,7 +368,7 @@ const Home = () => {
             </div>
             <TextContent
               type="black"
-              content="Piawood é o hub cultural da Casa Amarela, dedicado ao fomento da cultura e talentos locais na região de Barrinha, PI. Atuando como curador e plataforma de crescimento, Piawood promove eventos que integram música, artes visuais, gastronomia e esporte, gerando impacto na cena cultural da região. Além disso, a produtora Yellow Filme cria conteúdos audiovisuais que documentam e divulgam essas experiências, conectando a riqueza cultural local a um público mais amplo."
+              content="Piawood é o hub cultural da Casa Amarela, dedicado ao fomento da cultura e talentos locais na região de Barrinha, PI. Atuando como curador e plataforma de crescimento, Piawood promove eventos que integram música, artes visuais, gastronomia e esporte, gerando impacto na cena cultural da região. Além disso, a produtora Yellow Film cria conteúdos audiovisuais que documentam e divulgam essas experiências, conectando a riqueza cultural local a um público mais amplo."
               maxCharacters="maxCharacters30"
             />
           </div>
@@ -418,7 +418,7 @@ const Home = () => {
             <div className={styles.hospitalityInfo}>
               <div className={styles.hospitalityTitle}>
                 <Label
-                  label="Yellow Filme"
+                  label="Yellow Film"
                   type="orange"
                   icon={LogoLabelOrange}
                   border="squareRound"
@@ -426,7 +426,7 @@ const Home = () => {
               </div>
               <TextContent
                 type="blackBrown"
-                content="A Yellow Filme é a produtora audiovisual da Casa Amarela, focada em promover as atividades do hub cultural Piawood e do projeto de hospitalidade. A empresa atua em duas frentes: produção interna, criando conteúdos para divulgação e marketing, e serviços externos, oferecendo expertise local para marcas que buscam produções na região do Piauí. Com um foco regional, a Yellow Filme valoriza a cultura, os talentos e os cenários locais, criando conteúdos autênticos e de alto impacto visual."
+                content="A Yellow Film é a produtora audiovisual da Casa Amarela, focada em promover as atividades do hub cultural Piawood e do projeto de hospitalidade. A empresa atua em duas frentes: produção interna, criando conteúdos para divulgação e marketing, e serviços externos, oferecendo expertise local para marcas que buscam produções na região do Piauí. Com um foco regional, a Yellow Film valoriza a cultura, os talentos e os cenários locais, criando conteúdos autênticos e de alto impacto visual."
                 maxCharacters="maxCharacters30"
               />
               <Hyperlink
@@ -444,7 +444,7 @@ const Home = () => {
         id="videoBanner"
         className={`${styles.parallax} ${styles.videoBanner} ${styles.section}`}
       >
-        <iframe width="100%" height="900" src="https://www.youtube.com/embed/Hjq-ZzT1tjw?si=7ACuc3MS861aW3Uc" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" ></iframe>
+        <iframe width="100%" height="2000" src="https://www.youtube.com/embed/Hjq-ZzT1tjw?si=7ACuc3MS861aW3Uc" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" ></iframe>
         <Footer labelColor="yellowWhite" />
       </section>
     </div>
