+++ src/Components/Hospitality/Hospitality.module.css
@@ -13,6 +13,12 @@
   margin: 0 auto;
 }
 
+.image-gallery-content
+  .image-gallery-slide
+  .image-gallery-image {
+  height: 100px;
+}
+
 .container img {
   width: 100%;
 }
